requirements.txt
# Editable Git install with no remote (onenutil==1.0.0)
-e /Users/jm/repos/onenote-utils/src
azure-core==1.21.1
azure-identity==1.7.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
cryptography==36.0.1
elasticsearch==7.16.2
idna==3.3
joblib==1.1.0
msal==1.16.0
msal-extensions==0.3.1
msgraph-core==0.2.2
nltk==3.6.7
oauthlib==3.1.1
pdfminer.six==20211012
portalocker==2.3.2
prompt-toolkit==3.0.24
pycparser==2.21
pycryptodome==3.12.0
PyJWT==2.3.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
tqdm==4.62.3
urllib3==1.26.7
wcwidth==0.2.5
